AP_HAL_Linux: Fix truncation warnings

+++ libraries/AP_HAL_Linux/Storage.cpp
@@ -116,7 +116,7 @@ int Storage::_storage_create(const char *dpath)
     // take up all needed space
     if (ftruncate(fd, sizeof(_buffer)) == -1) {
         fprintf(stderr, "Failed to set file size to %u kB (%m)\n",
-                sizeof(_buffer) / 1024);
+                (unsigned int)(sizeof(_buffer) / 1024));
         goto fail;
     }
